+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Improve UX of course search pagination by avoiding truncation of page number
   when it is not relevant
 
+### Fixed
+
+- Fix unwanted comma when displaying course runs list on course detail page
+
 ## [2.0.1] - 2021-01-11
 
 ### Fixed